// server/admin/frontend/src/generated-client/model/linkPalletData.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Admin API
 * OpenAPI spec version: 1.0.0
 */
import type { EntryWithImage } from './entryWithImage';
import type { TwoHopLink } from './twoHopLink';

export interface LinkPalletData {
  /** Array of potential new link titles */
  newLinks: string[];
  /** Array of directly linked entries */
  links: EntryWithImage[];
  /** Array of two-hop link relationships */
  twohops: TwoHopLink[];
}